Домашняя работа 5 урока. Добавил еще одно активити. Приложение калькулятор готово!
