+++ part2/app/src/main/java/com/calculator/MainActivity.java
@@ -3,6 +3,7 @@ package com.calculator;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
@@ -13,113 +14,21 @@ import android.widget.Toast;
 
 public class MainActivity extends AppCompatActivity {
 
-    private final static String TAG = "MainActivity";
-    private TextView resultField; // текстовое поле для вывода результата
-    private TextView numberField;   // поле для ввода числа
-    private TextView operationField;    // текстовое поле для вывода знака операции
-    Double operand = null;  // операнд операции
-    String lastOperation = "="; // последняя операция
-
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        resultField = (TextView) findViewById(R.id.resultField);
-        numberField = (TextView) findViewById(R.id.numberField);
-        operationField = (TextView) findViewById(R.id.operationField);
+        initView();
 
     }
 
-    @Override
-    protected void onRestoreInstanceState(@NonNull Bundle savedInstanceState) {
-        lastOperation = savedInstanceState.getString("OPERATION");
-        operand = savedInstanceState.getDouble("OPERAND");
-        resultField.setText(operand.toString());
-        operationField.setText(lastOperation);
-        super.onRestoreInstanceState(savedInstanceState);
-        makeToast("onRestoreInstanceState");
-    }
-
-    @Override
-    protected void onSaveInstanceState(@NonNull Bundle outState) {
-        outState.putString("OPERATION", lastOperation);
-        if (operand != null)
-            outState.putDouble("OPERAND", operand);
-        super.onSaveInstanceState(outState);
-        makeToast("onSaveInstanceState");
-    }
-
-
-    // обработка нажатия на числовую кнопку
-    public void onNumberClick(View view) {
-
-        Button button = (Button) view;
-        numberField.append(button.getText());
-
-        if (lastOperation.equals("=") && operand != null) {
-            operand = null;
-        }
-    }
-
-    // обработка нажатия на кнопку операции
-    public void onOperationClick(View view) {
-
-        Button button = (Button) view;
-        String op = button.getText().toString();
-        String number = numberField.getText().toString();
-        // если введенно что-нибудь
-        if (number.length() > 0) {
-            number = number.replace(',', '.');
-            try {
-                performOperation(Double.valueOf(number), op);
-            } catch (NumberFormatException ex) {
-                numberField.setText("");
-            }
-        }
-        lastOperation = op;
-        operationField.setText(lastOperation);
-    }
-
-    private void performOperation(Double number, String operation) {
-
-        // если операнд ранее не был установлен (при вводе самой первой операции)
-        if (operand == null) {
-            operand = number;
-        } else {
-            if (lastOperation.equals("=")) {
-                lastOperation = operation;
-            }
-            switch (lastOperation) {
-                case "=":
-                    operand = number;
-                    break;
-                case "/":
-                    if (number == 0) {
-                        operand = 0.0;
-                    } else {
-                        operand /= number;
-                    }
-                    break;
-                case "*":
-                    operand *= number;
-                    break;
-                case "+":
-                    operand += number;
-                    break;
-                case "-":
-                    operand -= number;
-                    break;
-            }
-        }
-        resultField.setText(operand.toString().replace('.', ','));
-        numberField.setText("");
-    }
-
-    private void makeToast(String message) {
-        Toast.makeText(getApplicationContext(), message, Toast.LENGTH_SHORT).show();
-        Log.d(TAG, message);
+    private void initView() {
+        Button buttonStart = findViewById(R.id.start);
+        buttonStart.setOnClickListener(v -> {
+            Intent runActivity = new Intent(MainActivity.this, MainActivity2.class);
+            startActivity(runActivity);
+        });
     }
 }
