+++ part2/app/src/main/res/values/strings.xml
@@ -11,6 +11,7 @@
     <string name="_9">9</string>
     <string name="_plus">plus</string>
     <string name="_0">0</string>
+    <string name="Запустить">Запустить</string>
 
     <style name="ButtonStileCalculator">
         <item name="android:layout_margin">@dimen/button_margin</item>
